+++ src/editor/interface/Area.ts
@@ -1,9 +1,11 @@
 import { AreaMode } from '../dataset/enum/Area'
 import { LocationPosition } from '../dataset/enum/Common'
 import { IElement, IElementPosition } from './Element'
+import { IPlaceholder } from './Placeholder'
 
 export interface IAreaBasic {
   extension?: unknown
+  placeholder?: IPlaceholder
 }
 
 export interface IAreaStyle {
@@ -14,6 +16,7 @@ export interface IAreaStyle {
 
 export interface IAreaRule {
   mode?: AreaMode
+  deletable?: boolean
 }
 
 export type IArea = IAreaBasic & IAreaStyle & IAreaRule