+++ src/editor/interface/Control.ts
@@ -107,6 +107,7 @@ export interface IControlOption {
   borderWidth?: number
   borderColor?: string
   activeBackgroundColor?: string
+  disabledBackgroundColor?: string
 }
 
 export interface IControlInitOption {
@@ -141,6 +142,7 @@ export interface IControlContext {
 
 export interface IControlRuleOption {
   isIgnoreDisabledRule?: boolean // 忽略禁用校验规则
+  isIgnoreDeletedRule?: boolean // 忽略删除校验规则
   isAddPlaceholder?: boolean // 是否添加占位符
 }